아이템 능력치 int 값 float로 변경

+++ SpartaDungeon/Scripts/Player/Player.cs
@@ -136,13 +136,13 @@ namespace SpartaDungeon
                     switch (item.Stat)
                     {
                         case Stat.Health:
-                            BonusFullHP += item.StatValue;
+                            BonusFullHP += (int)item.StatValue;
                             break;
                         case Stat.Attack:
-                            BonusAttack += item.StatValue;
+                            BonusAttack += (int)item.StatValue;
                             break;
                         case Stat.Defense:
-                            BonusDefense += item.StatValue;
+                            BonusDefense += (int)item.StatValue;
                             break;
                         default:
                             break;
